move the v1 ReadConfig to v1 package
