+++ Makefile
@@ -45,7 +45,7 @@ coverage: ## run the go coverage tool, reading file coverage.out
 	go tool cover -html=coverage.out
 
 test: 
-	gotest -race -coverprofile=coverage.txt -covermode=atomic ./...
+	gotest -race ./...
 
 test-coverage: ## run the test with proper coverage reporting
 	goverage -race -coverprofile=coverage.out -covermode=atomic ./...
